requirements/dev.txt
wemake-python-styleguide==0.15.3
flake8==3.9.2
black==21.9b0
isort==5.10.1
mypy==0.910
Django==3.2
djangorestframework==3.13.1
django-stubs==1.9.0
django-stubs-ext==0.3.1
djangorestframework-stubs
